+++ src-gui/src/renderer/components/pages/help/SettingsPage.tsx
@@ -4,6 +4,7 @@ import DonateInfoBox from "./DonateInfoBox";
 import DaemonControlBox from "./DaemonControlBox";
 import SettingsBox from "./SettingsBox";
 import ExportDataBox from "./ExportDataBox";
+import DiscoveryBox from "./DiscoveryBox";
 import { useLocation } from "react-router-dom";
 import { useEffect } from "react";
 
@@ -27,6 +28,7 @@ export default function SettingsPage() {
       }}
     >
       <SettingsBox />
+      <DiscoveryBox />
       <ExportDataBox />
       <DaemonControlBox />
       <DonateInfoBox />