+++ swap/src/bitcoin/electrum_balancer.rs
@@ -263,15 +263,15 @@ where
             let idx = self.next.load(Ordering::SeqCst);
 
             // Get client for this index
-            let client = self.get_or_init_client_sync(idx).map_err(|e| {
+            let client = self.get_or_init_client_sync(idx).map_err(|err| {
                 trace!(
                     server_url = self.urls[idx],
                     attempt = errors.len(),
-                    error = ?e,
+                    error = ?err,
                     "Client initialization failed, switching to next client"
                 );
 
-                errors.push(e);
+                errors.push(err);
 
                 BackoffError::transient(())
             })?;